rename identification_view.dart to identification_page.dart to match class name
